[AIRFLOW-6721] Move WebHdfsSensor tests to own module
Move the WebHdfsSensor tests out of the operator tests and into a
separate module.
